Add line endings to non-PHP files
I have only been dealing with PHP files in this branch so far, so the non-PHP files
still lack line endings. This commit adds line endings to only those files, which
should ensure that they are in line with the files on `main` and do not appear in
this PR's diff view.