docs: amend code of conduct
<opinion>
Explicit enumeration makes sense in legal documents, but CoCs are
more aspirational so I don't think it makes sense in this context.
You will end up either having to constantly add new things to the
list, or declare that the list is non-exhaustive (in which there's
not much point in enumerating in the first place).

This is doubly true given the inherently woolly realm of appropriate
interpersonal interactions that a CoC is implemented to moderate.

Also, remember the KISS principle. Personally, I think this is the
pinnacle of CoC design: https://github.com/karlgroves/dontbeadick
However, I'm sure there is a slightly-more-professional middle
ground ;)
</opinion>